+++ CHANGELOG.md
@@ -19,6 +19,7 @@
   - Added `TracePipeStdout()`
   - Lists now write status code and error message when tracing enabled
   - Pipelines now write status code and error message when tracing enabled
+  - `Chmod()` now supports tracing
 
 ## v1.1.1